requirements.txt
pytz==2024.1
Flask==3.0.3
gunicorn==23.0.0
selenium==4.21.0
google-cloud-secret-manager==2.20.0
google-cloud-firestore==2.17.2
google-auth==2.34.0
firebase-admin==6.5.0
